Removed optional Jupyter dependency and refactored dependency checking logic into its own sub-package
