+++ src/uv_secure/__init__.py
@@ -1,5 +1,5 @@
 from uv_secure.__version__ import __version__
-from uv_secure.run import app, check_dependencies
+from uv_secure.run import app
 
 
-__all__ = ["__version__", "app", "check_dependencies"]
+__all__ = ["__version__", "app"]
