+++ src/uv_secure/dependency_checker/__init__.py
@@ -1,0 +1,4 @@
+from uv_secure.dependency_checker.dependency_checker import check_dependencies
+
+
+__all__ = ["check_dependencies"]
